+++ compiler/middleend/CheckerContext.ts
@@ -32,6 +32,7 @@ export interface CheckerContext {
   // Type Resolution & Compatibility
   resolveType(type: AST.TypeNode, checkConstraints?: boolean): AST.TypeNode;
   ensureKnownType(type: AST.TypeNode): void;
+  ensureImplicitPrimitiveWrappersLoaded(typeName: string): void;
   areTypesCompatible(
     t1: AST.TypeNode,
     t2: AST.TypeNode,